cast on the return from Json

+++ news/lib/src/resources/news_api_provider.dart
@@ -9,7 +9,7 @@ Client client = Client();
   Future<List<int>>fetchTopIds() async {
   final response = await client.get('$_root/topstories.json');
   final ids = json.decode(response.body);
-  return ids;
+  return ids.cast<int>();
   }
 
   Future<ItemModel>fetchItem(int id)async {
